fix: fix issue with memoized token announcements
