+++ .changeset/cuddly-mangos-hear.md
@@ -1,0 +1,5 @@
+---
+"@venusprotocol/evm": patch
+---
+
+fix issue with memoized token announcements
